backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
fastapi==0.95.1
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pydantic==1.10.7
pytz==2022.7.1
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
